MappedRegion: Just use uint64_t for all memory addresses.
It doesn't matter that those values "point" to memory in some way -
the memory is in another process and we're certainly not going to
dereference the "pointers".

+++ pageinfo.h
@@ -12,8 +12,8 @@
 
 struct MappedRegion
 {
-    uintptr_t start;
-    uintptr_t end;
+    uint64_t start;
+    uint64_t end;
     std::vector<uint32_t> useCounts;
     std::vector<uint32_t> combinedFlags;
     bool operator<(const MappedRegion &other) const { return start < other.start; }
